backend/requirements.txt
Flask
flask-Cors
python-dotenv
google-cloud-aiplatform